Refactor test directory structure and imports for clarity and consistency
- Renamed test directory  to .
- Updated import paths in affected test files to reflect the new directory structure.
- Removed obsolete  file.
