+++ tests/test_accept_string/test_smiles.py
