+++ tests/test_accept_token_sequence/_test_accept_tokens_mixin.py
