+++ tests/test_accept_token_sequence/test_bloom.py
@@ -2,7 +2,9 @@ import unittest
 
 from transformers import BloomTokenizerFast
 
-from tests._test_token_seq_recognizer_many_tokenizer_common import TokenizerTesterMixin
+from tests.test_accept_token_sequence._test_accept_tokens_mixin import (
+    TokenizerTesterMixin,
+)
 
 
 @unittest.skip("Bloom is not supported and will be removed")
